pyproject.toml
[project]
name = "pipecat-sonos-voice-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "mcp[cli]",
  "aiohttp",
  "pipecat-ai[webrtc,deepgram,cartesia,mcp,openai]",
  "onnxruntime",
  "pipecat-ai-small-webrtc-prebuilt",
  "mcp-registry",
  "sonos-mcp-server @ git+https://github.com/WinstonFassett/sonos-mcp-server.git@dev"
]
